fixes flaky input tests by manually triggering keydown events (#321)
inputs are weird. inputs in test environments are weirder. this forces
a keydown event on the inputs so the compnent runs its handler and
updates the proper props.

I think that `keyUp` is technically the proper event to use in the
component, but let's get a solid pattern first and think about that
refactor later, if at all.